Update heroku-tailscale-start.sh
Use a consistent socket location otherwise the existence of a /var/run directory confuses tailscale. The root issue is that Heroku's internal buildpack-built dynos do not have a /var/run, while the Heroku CNB build environment creates a /var/run. It would be helpful for Heroku to publish the actual buildpacks being used if they intend to support the CNB.